revert structure
